% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajemv.R
\name{ajemv}
\alias{ajemv}
\title{Maximum likelihood fit}
\usage{
ajemv(dados, est = T, pepita = 0, ini.phi, ini.sigma2,
  modelo = "exponential", restrita = F, plot = T, col = 3, boxcox = F,
  ini.lambda = 0.5, tendencia = F, tend = "1st")
}
\arguments{
\item{dados}{*}

\item{est}{*}

\item{pepita}{*}

\item{ini.phi}{*}

\item{ini.sigma2}{*}

\item{modelo}{*}

\item{restrita}{*}

\item{plot}{*}

\item{col}{*}

\item{boxcox}{*}

\item{ini.lambda}{*}

\item{tendencia}{*}

\item{tend}{*}
}
\value{
The fitted model
}
\description{
Fits the spatial covariance model through maximum likelihood
}
