% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajmqp.R
\name{ajmqp}
\alias{ajmqp}
\title{Minimum squares fit}
\usage{
ajmqp(dados, est = T, pepita = 0, ini.phi, ini.sigma2,
  modelo = "exponential", pesos = "npairs", ptos = 15, plot = T,
  col = 2)
}
\arguments{
\item{dados}{*}

\item{est}{*}

\item{pepita}{*}

\item{ini.phi}{*}

\item{ini.sigma2}{*}

\item{modelo}{*}

\item{pesos}{*}

\item{ptos}{*}

\item{plot}{*}

\item{col}{*}
}
\value{
The fitted model
}
\description{
Fits the spatial covariance model through minimum squares
}
