% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Inserts an element in a vector}
\usage{
insert(vec, el, pos)
}
\arguments{
\item{vec}{Vector of interest}

\item{el}{Element to be inserted}

\item{pos}{Position to be inserted. Must be lesser or equal to the length of
`vec`}
}
\value{
A new vector based on `vec` with the new element
}
\description{
Inserts an element `el` into the vector `vec` in the position `pos`
}
