% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvar.R
\name{plotvar}
\alias{plotvar}
\title{Plots the robust sample variogram}
\usage{
plotvar(dados, ptos)
}
\arguments{
\item{dados}{Georeferenced data, in the form of a geodata object}

\item{ptos}{Number of points in the variogram}
}
\value{
Robust variogram
}
\description{
Plots the robust sample variogram
}
